Add files via upload
an exploratory data analysis on the Cleveland heart disease dataset, trained three types of classifiers to diagnose heart disease based on the available data and compare them visually in a grouped bar chart with regard to accuracy, sensitivity and specificity. Create a confusion matrix for each classifier and plot them as a heatmap.